Typo fixes, extra class TestGuzzleHttpClient deleted, arguments order changed
